pyproject.toml
[tool.poetry]
name = "complaints-test-app"
version = "0.1.0"
description = ""
authors = ["Alexander <sashamorozov97@mail.ru.>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
uvicorn = "^0.35.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
aiosqlite = "^0.21.0"
httpx = "^0.28.1"
pydantic-settings = "^2.10.1"
alembic = "^1.16.4"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
